rename to 'putViewColumnCell'
